Rewrite and Restructure
Kept on crashing out to how I was writing code before where I kept having to completly rewrite code that would break and just not do what I desire. One of the issues was the having multiple diffrent client and server scripts which then problems to do with remote events would occur. So now code is restructured and will be writin with ONE server script and ONE client script that will handle the entire game, and then have module scripts childed to each of them for OOP classes or any other abstracted code that a programmer may desire.